+++ frontend/lib/presentation/screens/search/on_search.dart
@@ -1,0 +1,36 @@
+
+
+import 'package:flutter/material.dart';
+import 'package:frontend/presentation/screens/search/small_widget_search.dart';
+import 'package:frontend/presentation/screens/search/wide_widget_search.dart';
+
+import 'package:frontend/presentation/utils/screen_adjuster.dart';
+
+class OnSearchScreen extends StatefulWidget {
+  final Function(Locale) onLocaleChange;
+
+  const OnSearchScreen({
+    super.key, 
+    required this.onLocaleChange
+    });
+
+  @override
+  OnSearchScreenState createState() => OnSearchScreenState();
+}
+
+class OnSearchScreenState extends State<OnSearchScreen>  {
+
+  @override
+  Widget build(BuildContext context) {
+
+    final adjuster = ScreenAdjuster<Widget>(
+      smallWidget: SmallWidgetSearch(onLocaleChange: widget.onLocaleChange),
+      wideWidget: const WideWidgetSearch(),
+      threshold: 800, // Customize threshold for responsiveness
+    );
+    
+    return adjuster.adjust(context);
+
+  }
+
+}